Pulse: Navbar left in landscape support
By default now, AOSP places the navigation bar on the left side
of the screen on phones when rotated 270 degrees. Adapt for this
change by updating Pulse when orientation changes.

The orientation listener will be enabled disabled according to the screen
state (screen on, screen off, keyguard showing or not, keyguard bypassed after
fingerprint unlock thus screen on straight).

Change-Id: Id9b8e43baca11a314cc137940852b9842069b77f